Cleanup code structure
