+++ src/modules/component/episodeListItem/episodeListItem.js
@@ -1,6 +1,6 @@
 import { LightningElement, api, track, wire } from 'lwc';
 
-import { connectStore, store } from 'store/store';
+import { connectStore, store } from 'rcast/store';
 
 import { convertMilliseconds, formatDate } from 'base/utils';
