+++ src/modules/component/player/player.js
@@ -1,8 +1,8 @@
 import { LightningElement, api, track, wire } from 'lwc';
 
-import player from 'store/player';
-import { connectStore, store } from 'store/store';
-import { play, pause } from 'store/actions';
+import player from 'rcast/player';
+import { connectStore, store } from 'rcast/store';
+import { play, pause } from 'rcast/store';
 
 const STEP_DURATION = 15;
