+++ src/modules/rcast/player/event-emitter.js
