+++ src/modules/rcast/store/constants.js
@@ -1,0 +1,2 @@
+export const RECORD_TYPE_HIGHLIGHT = 'HIGHLIGHT';
+export const RECORD_TYPE_FULL = 'FULL';
