+++ src/modules/rcast/store/modules/application/actions.js
@@ -1,0 +1,8 @@
+import { CONNECTIVITY_STATUS_CHANGED } from './constants';
+
+export function connectivityStatusChanged() {
+    return {
+        type: CONNECTIVITY_STATUS_CHANGED,
+        isOnline: navigator.onLine,
+    };
+}
