+++ src/modules/rcast/store/modules/application/constants.js
@@ -1,0 +1,1 @@
+export const CONNECTIVITY_STATUS_CHANGED = 'CONNECTIVITY_STATUS_CHANGED';
