+++ src/modules/rcast/store/modules/categories/constants.js
@@ -1,0 +1,3 @@
+export const REQUEST_CATEGORY = 'REQUEST_CATEGORY';
+export const RECEIVE_CATEGORY_SUCCESS = 'RECEIVE_CATEGORY_SUCCESS';
+export const RECEIVE_CATEGORY_ERROR = 'RECEIVE_CATEGORY_ERROR';
