+++ src/modules/rcast/store/modules/episodes/constants.js
@@ -1,0 +1,3 @@
+export const REQUEST_EPISODE = 'REQUEST_EPISODE';
+export const RECEIVE_EPISODE_SUCCESS = 'RECEIVE_EPISODE_SUCCESS';
+export const RECEIVE_EPISODE_ERROR = 'RECEIVE_EPISODE_ERROR';
