+++ src/modules/rcast/store/modules/info/constants.js
@@ -1,0 +1,6 @@
+export const SUBSCRIBE_SHOW = 'SUBSCRIBE_SHOW';
+export const UNSUBSCRIBE_SHOW = 'UNSUBSCRIBE_SHOW';
+export const DOWNLOAD_EPISODE_PROGRESS = 'DOWNLOAD_EPISODE_PROGRESS';
+export const DOWNLOAD_EPISODE_DONE = 'DOWNLOAD_EPISODE_DONE';
+export const DOWNLOAD_EPISODE_ERROR = 'DOWNLOAD_EPISODE_ERROR';
+export const DISCARD_DOWNLOADED_EPISODE = 'DISCARD_DOWNLOADED_EPISODE';
