+++ src/modules/rcast/store/modules/player/constants.js
@@ -1,0 +1,4 @@
+export const PLAY = 'PLAY';
+export const PAUSE = 'PAUSE';
+export const ENDED = 'ENDED';
+export const LISTEN_EPISODE = 'LISTEN_EPISODE';
