+++ src/modules/rcast/store/modules/shows/constants.js
@@ -1,0 +1,3 @@
+export const REQUEST_SHOW = 'REQUEST_SHOW';
+export const RECEIVE_SHOW_SUCCESS = 'RECEIVE_SHOW_SUCCESS';
+export const RECEIVE_SHOW_ERROR = 'RECEIVE_SHOW_ERROR';
