--- src/modules/store/shared/shared.js
@@ -1,31 +1,0 @@
-export const CONNECTIVITY_STATUS_CHANGED = 'CONNECTIVITY_STATUS_CHANGED';
-
-export const REQUEST_SHOW = 'REQUEST_SHOW';
-export const RECEIVE_SHOW_SUCCESS = 'RECEIVE_SHOW_SUCCESS';
-export const RECEIVE_SHOW_ERROR = 'RECEIVE_SHOW_ERROR';
-
-export const REQUEST_EPISODE = 'REQUEST_EPISODE';
-export const RECEIVE_EPISODE_SUCCESS = 'RECEIVE_EPISODE_SUCCESS';
-export const RECEIVE_EPISODE_ERROR = 'RECEIVE_EPISODE_ERROR';
-
-export const REQUEST_CATEGORY = 'REQUEST_CATEGORY';
-export const RECEIVE_CATEGORY_SUCCESS = 'RECEIVE_CATEGORY_SUCCESS';
-export const RECEIVE_CATEGORY_ERROR = 'RECEIVE_CATEGORY_ERROR';
-
-export const SUBSCRIBE_PODCAST = 'SUBSCRIBE_PODCAST';
-export const UNSUBSCRIBE_PODCAST = 'UNSUBSCRIBE_PODCAST';
-
-export const PLAY = 'PLAY';
-export const PAUSE = 'PAUSE';
-export const ENDED = 'ENDED';
-
-export const LISTEN_EPISODE = 'LISTEN_EPISODE';
-
-export const DOWNLOAD_EPISODE_PROGRESS = 'DOWNLOAD_EPISODE_PROGRESS';
-export const DOWNLOAD_EPISODE_DONE = 'DOWNLOAD_EPISODE_DONE';
-export const DOWNLOAD_EPISODE_ERROR = 'DOWNLOAD_EPISODE_ERROR';
-
-export const DISCARD_DOWNLOADED_EPISODE = 'DISCARD_DOWNLOADED_EPISODE';
-
-export const RECORD_TYPE_HIGHLIGHT = 'HIGHLIGHT';
-export const RECORD_TYPE_FULL = 'FULL';
